Create API (year, scoreboard....)

+++ SupportRegister.Data/Models/Application.cs
@@ -13,10 +13,10 @@ namespace SupportRegister.Data.Models
         }
 
         public int IdApplication { get; set; }
-        public string NameApplication { get; set; }
+        public string NameApplication { get; set; }    
         public string Description { get; set; }
         public string Content { get; set; }
-        public decimal? Price { get; set; }
+        public int? Price { get; set; }
 
         public ICollection<DetailRegisterApplication> DetailRegisterApplications { get; set; }
     }
