+++ SupportRegister.Data/Models/DetailRegisterApplication.cs
@@ -10,7 +10,7 @@ namespace SupportRegister.Data.Models
         public int IdApplication { get; set; }
         public int IdRegisterApplication { get; set; }
         public int? Amount { get; set; }
-        public decimal? Price { get; set; }
+        public int? Price { get; set; }
 
         public Application IdApplicationNavigation { get; set; }
         public RegisterApplication IdRegisterApplicationNavigation { get; set; }
